[Mosaic] Extract Mosaic GameScreen composable in its own file
